requirements.txt
-e git+https://github.com/django/django.git#egg=django

south
django-extensions


ipython
ipdb

psycopg2
dateutils
lxml
pyyaml

django-tastypie
django-tastypie-swagger

gunicorn

pylint
pep8

nose
django-nose
tox

dj-database-url
django-postgrespool

django-haystack
whoosh==2.4

